Create Manager servlet and JSP
Added a fake staff.jsp to hold the place for the login later, fixed
WelcomeServlet message which was probably bothering me in my version of
Java.
Put some things in the Booking bean, just to pass between the servlet
and the jsp.
Set up the manager servlet to pass on some fake booking data, and made
manager.jsp display it. All still WIP...