Refine managing and assigning bookings
(This is a bunch of little stuff, trying to check it in aside from the
next couple of bigger changes, the room type bean and the staff login.
This is approaching all I can do until we've finalised the room status
stuff...)
Rearranged manager bookings into two sorts, filled and unfilled, adding
in the proper data link to assigning a room.
Put a bit more logic into the assigning-room servlet, but I'll add the
data structure in a separate commit.
Also added bits to the Booking bean; more to come!