Add TypeRequest, beginning to use it
Added a TypeRequest class for passing to the jsp. It stores a room type,
number of rooms requested, and the list of available rooms.
Modified the servlet to use the type and think about when else to use
it.
Initial test with the jsp is in there, more to come of course but it
works right now.
Minor fix on manager.jsp