Added login security checks and logout buttons
OK. Staff login page should be working.
Log in as owner, get a session attribute for it.
Or log in as manager, get a session attribute for that.
Then get redirected to the relevant home page. If you try to go back to
/staff, it keeps redirecting you to your home page.
If you try to go to one of the staff pages *without* being logged in, it
sends you to /staff to log in. This is true for my /manage and
/assignrooms, and the owner's /availability, /offer/ and /maintenance,
or whatever those three servlets were called. If we want to add it to
others, it's a few easy if/else lines at the top of any doGet or doPost.
I've added a log out button - which works for owners or managers - to
the bottom of /manage and /assignrooms, and /availability. Let me know
if the other two owner pages should have it too?